requirements.txt
aiobotocore==2.5.4
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioitertools==0.12.0
aiosignal==1.3.1
anndata==0.11.1
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.9.1
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
botocore==1.31.17
cellpose==3.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dask==2024.12.0
dask-expr==1.1.20
dask-image==2024.5.3
datashader==0.16.3
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.12.0
docrep==0.3.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.21.1
fastremap==1.15.0
filelock==3.13.1
fonttools==4.55.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.6.0
fuzzy-c-means==1.7.2
geopandas==1.0.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
igraph==0.11.8
imagecodecs==2024.9.22
imageio==2.36.1
importlib_metadata==8.5.0
inflect==7.4.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lime==0.2.0.1
llvmlite==0.43.0
locket==1.0.0
mahotas==1.4.18
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multiscale_spatial_image==1.0.1
natsort==8.4.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.1
notebook_shim==0.2.4
numba==0.60.0
numcodecs==0.14.1
numpy==1.26.3
ome-zarr==0.9.0
omnipath==1.0.8
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
param==2.1.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pillow==10.2.0
PIMS==0.7
platformdirs==4.3.6
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pyct==0.5.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pynndescent==0.5.13
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roifile==2024.9.15
rpds-py==0.22.3
s3fs==2023.6.0
scanpy==1.10.4
scikit-fuzzy==0.5.0
scikit-image==0.25.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
session_info==1.0.0
shapely==2.0.6
six==1.17.0
slicerator==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spatial_image==1.1.0
spatialdata==0.2.5.post0
squidpy==1.6.2
stack-data==0.6.3
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
tabulate==0.8.10
tblib==3.0.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
toolz==1.0.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.1
typer==0.9.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==1.26.20
validators==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.0
xarray==2024.7.0
xarray-dataclasses==1.9.1
xarray-datatree==0.0.15
xarray-schema==0.0.3
xarray-spatial==0.4.0
yarl==1.18.3
zarr==2.18.3
zict==3.0.0
zipp==3.21.0
